Enable custom identity entities, enable email verification

+++ BlockBot.Web/Controllers/ProjectsController.cs
@@ -2,17 +2,22 @@
 using System.Threading.Tasks;
 using BlockBot.Web.Data;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Query;
 
 namespace BlockBot.Web.Controllers
 {
     public class ProjectsController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly ApplicationUserManager _userManager;
 
-        public ProjectsController(ApplicationDbContext context)
+        public ProjectsController(ApplicationDbContext context,
+            ApplicationUserManager userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         // GET: Projects/Details/5
@@ -23,11 +28,23 @@ namespace BlockBot.Web.Controllers
                 return NotFound();
             }
 
-            var project = await _context.Projects
+            Project project = await _context.Projects
+                .Include(p => p.Owner)
                 .FirstOrDefaultAsync(m => m.ProjectId == id);
             if (project == null)
             {
                 return NotFound();
+            }
+
+            ApplicationUser user = await _userManager.GetUserAsync(User);
+            if (user == null)
+            {
+                return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
+            }
+
+            if (project.OwnerId != user.Id)
+            {
+                return Unauthorized();
             }
 
             return View(project);
@@ -44,23 +61,25 @@ namespace BlockBot.Web.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ProjectId,Name")] Project project)
+        public async Task<IActionResult> Create([Bind("Name")] Project project)
         {
             if (ModelState.IsValid)
             {
-                _context.Add(project);
-                await _context.SaveChangesAsync();
-
-                _context.ProjectSteps.Add(new ProjectStep
+                ApplicationUser user = await _userManager.GetUserAsync(User);
+                if (user == null)
                 {
-                    ProjectId = project.ProjectId,
-                    ProjectStepOrder = 1,
-                    StepXML = "<xml xmlns=\"http://www.w3.org/1999/xhtml\" id=\"workspaceBlocks\" style=\"display: none;\"></xml>"
-                });
-                await _context.SaveChangesAsync();
+                    return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
+                }
 
+                project.OwnerId = user.Id;
+                project.XML =
+                    "<xml xmlns=\"http://www.w3.org/1999/xhtml\" id=\"workspaceBlocks\" style=\"display: none;\"></xml>";
+
+                _context.Add(project);
+                    await _context.SaveChangesAsync();
                 return RedirectToAction("Dashboard", "Dashboard");
             }
+
             return View(project);
         }
 
@@ -72,11 +91,23 @@ namespace BlockBot.Web.Controllers
                 return NotFound();
             }
 
-            var project = await _context.Projects.FindAsync(id);
+            Project project = await _context.Projects.FindAsync(id);
             if (project == null)
             {
                 return NotFound();
             }
+
+            ApplicationUser user = await _userManager.GetUserAsync(User);
+            if (user == null)
+            {
+                return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
+            }
+
+            if (project.OwnerId != user.Id)
+            {
+                return Unauthorized();
+            }
+            
             return View(project);
         }
 
@@ -92,6 +123,24 @@ namespace BlockBot.Web.Controllers
                 return NotFound();
             }
 
+            ApplicationUser user = await _userManager.GetUserAsync(User);
+            if (user == null)
+            {
+                return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
+            }
+
+            Project originalProject = await _context.Projects.FindAsync(id);
+
+            if (originalProject == null) 
+            {
+                return NotFound();
+            }
+
+            if (originalProject.OwnerId != user.Id)
+            {
+                return Unauthorized();
+            }
+
             if (ModelState.IsValid)
             {
                 try
@@ -105,13 +154,13 @@ namespace BlockBot.Web.Controllers
                     {
                         return NotFound();
                     }
-                    else
-                    {
-                        throw;
-                    }
+
+                    throw;
                 }
+
                 return RedirectToAction("Dashboard", "Dashboard");
             }
+
             return View(project);
         }
 
@@ -123,24 +172,52 @@ namespace BlockBot.Web.Controllers
                 return NotFound();
             }
 
-            var project = await _context.Projects
+            Project project = await _context.Projects
+                .Include(p => p.Owner)
                 .FirstOrDefaultAsync(m => m.ProjectId == id);
             if (project == null)
             {
                 return NotFound();
             }
 
+            ApplicationUser user = await _userManager.GetUserAsync(User);
+            if (user == null)
+            {
+                return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
+            }
+
+            if (project.OwnerId != user.Id)
+            {
+                return Unauthorized();
+            }
+
             return View(project);
         }
 
         // POST: Projects/Delete/5
-        [HttpPost, ActionName("Delete")]
+        [HttpPost]
+        [ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var project = await _context.Projects.FindAsync(id);
-            _context.ProjectSteps.RemoveRange(
-                _context.ProjectSteps.Where(x => x.ProjectId == project.ProjectId));
+            Project project = await _context.Projects.FindAsync(id);
+
+            if (project == null)
+            {
+                return NotFound();
+            }
+
+            ApplicationUser user = await _userManager.GetUserAsync(User);
+            if (user == null)
+            {
+                return NotFound($"Unable to load user with ID '{_userManager.GetUserId(User)}'.");
+            }
+
+            if (project.OwnerId != user.Id)
+            {
+                return Unauthorized();
+            }
+
             _context.Projects.Remove(project);
             await _context.SaveChangesAsync();
             return RedirectToAction("Dashboard", "Dashboard");
